requirements.txt
fastapi
pydantic
transformers
accelerate
flask
langchain
langchain-groq
chromadb
python-dotenv
PyPDF2
bcrypt
gunicorn
Flask-Cors
sentence-transformers
python-multipart
pysqlite3-binary
git+https://github.com/openai/whisper.git
moviepy